14 oct

+++ lib/view/Invoices/new_invoice.dart
@@ -802,7 +802,7 @@ class _NewInvoiceState extends State<NewInvoice> {
                                                                       invoiceController
                                                                               .invoiceItems[
                                                                           index],
-                                                                      double.tryParse(
+                                                                      int.tryParse(
                                                                           New_Qty
                                                                               .text)!);
                                                                   Navigator.of(
